clarifications on strings
IMO the sentences in note contained unnecessary complication, as strings are always collections (sequences) of characters , regardless of encoding. 
Removed "Therefore" before "it is immutable and cannot be indexed", because being UTF-8 does not imply those, strictly speaking.
Also when I ran the Ist example, 
'Hello' at: 1.  
my Cuis produced $H(000048) , not just $H ,so edited the ex accordingly.

And a question: why say that it cannot be indexed, if (the new UTF8) strings can still receive the message 'at: ' and 'indexOf', even without using 'asByteString' ? 
Perhaps it will not work correctly for all strings? If so, maybe say "not all strings will be indexed correctly"? And specify indexing will work correctly if the string is made up of ASCII characters , I guess ?...

But in general, concerning the language: I believe all this is internal implementation detail that should not concern the users of Cuis. Or else, it goes against the simplicity and cleanliness tradition of Smalltalk, no? 
The new strings should be able to correctly receive #at and  #indeOf , as well as be changed using 'at: put:' etc, without having to be converted to byte strings first (or at least not by the user, explicitely). 
Having to use #asBytestring so often is a nuisance.